requirements.txt
asgiref==3.11.0
attrs==25.4.0
certifi==2026.1.4
charset-normalizer==3.4.4
cloudinary==1.44.1
Django==6.0.1
django-cloudinary-storage==0.3.0
djangorestframework==3.16.1
drf-spectacular==0.29.0
et_xmlfile==2.0.0
flake8==7.3.0
idna==3.11
inflection==0.5.1
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
mccabe==0.7.0
openpyxl==3.1.5
pillow==12.1.0
pycodestyle==2.14.0
pyflakes==3.4.0
python-dotenv==1.2.1
PyYAML==6.0.3
referencing==0.37.0
requests==2.32.5
rpds-py==0.30.0
six==1.17.0
sqlparse==0.5.5
uritemplate==4.2.0
urllib3==2.6.3
